refactor: move ExecutionStatus to pkg/execution
Extract ExecutionStatus enum from pkg/types to pkg/execution as it's
specific to the execution context, not a universal type. Update all
references to use the new package.

Part of #704
